+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [#145] Update `item_shipped` in LenderActions to process and apply dueDate returned by Rapido API, updating both checkout and ILL request due dates with buffer subtracted
+- [#145] Refactor LenderActions.t to use mocked plugin pattern for consistency with BorrowerActions.t
 
 ## [1.4.0] - 2026-02-04